requirements.txt
requests>=2.31.0
pandas>=2.0.3
numpy>=1.24.3
tensorflow>=2.12.0
scikit-learn>=1.3.0
python-telegram-bot==13.15
python-dotenv>=1.0.0